first test for simple file modification detection (#470)
